fix(cockpit): Move getRoleList and updateUserRoles to role-actions, rename getUserList
