+++ apps/cockpit/src/lib/role-actions.ts
@@ -4,10 +4,116 @@ import type { TRoleDetailFormSchema } from "@/lib/rbac-schema";
 import type { TCreateRoleFormData } from "@/lib/rbac-utils";
 import { db } from "@northware/database/connection";
 import { handleNeonError } from "@northware/database/neon-error-handling";
-import { permissionsToRoles, rolesTable } from "@northware/database/schema";
+import {
+  permissionsTable,
+  permissionsToRoles,
+  rolesTable,
+  rolesToAccounts,
+} from "@northware/database/schema";
 import { and, eq, inArray } from "drizzle-orm";
 import { revalidatePath } from "next/cache";
 import { cache } from "react";
+import type {
+  TRoleListResponse,
+  TRoleWithPermissions,
+  TUpdateRolesParams,
+} from "./rbac-types";
+
+export async function getRoleList(): Promise<TRoleListResponse> {
+  try {
+    const response = await db
+      .select({
+        recordId: rolesTable.recordId,
+        roleKey: rolesTable.roleKey,
+        roleName: rolesTable.roleName,
+        permissionKey: permissionsTable.permissionKey,
+        permissionName: permissionsTable.permissionName,
+      })
+      .from(rolesTable)
+      .leftJoin(
+        permissionsToRoles,
+        eq(rolesTable.roleKey, permissionsToRoles.roleKey)
+      )
+      .leftJoin(
+        permissionsTable,
+        eq(permissionsToRoles.permissionKey, permissionsTable.permissionKey)
+      );
+
+    const result: Record<string, TRoleWithPermissions> = {};
+    for (const item of response) {
+      if (!result[item.roleKey]) {
+        result[item.roleKey] = {
+          recordId: item.recordId,
+          roleKey: item.roleKey,
+          roleName: item.roleName,
+          permissions: [],
+        };
+      }
+
+      if (item.permissionKey !== null) {
+        result[item.roleKey].permissions.push({
+          permissionKey: item.permissionKey,
+          permissionName: item.permissionName,
+        });
+      }
+    }
+    return { success: true, roleList: Object.values(result) };
+  } catch (error) {
+    return {
+      success: false,
+      error: error instanceof Error ? error : new Error("Unknown Error"),
+    };
+  }
+}
+export async function updateUserRoles({
+  data,
+  userRolesResponse,
+  userId,
+}: TUpdateRolesParams) {
+  // filtert aus den übergebenen Formulardaten die roleKeys der aktiven Switches heraus
+  const selectedRoles = Object.entries(data)
+    .filter(([_, value]) => value) // Nur ausgewählte Rollen (value === true)
+    .map(([roleKey]) => roleKey); // Extrahiere die roleKeys
+
+  // enthält roleKeys, die in selecctedRoles aber nicht in userRolesResponse enthalten sind
+  const rolesToAdd = selectedRoles.filter(
+    (selectedRole) => !userRolesResponse.includes(selectedRole)
+  );
+  // enthält roleKeys, die in userRolesRespnse aber nicht in selectedRoles enthalten sind
+  const rolesToRemove = userRolesResponse
+    .filter((userRole): userRole is string => userRole !== null)
+    .filter((userRole) => !selectedRoles.includes(userRole));
+
+  const insertRoles = new Array();
+  rolesToAdd.forEach((role, i) => {
+    insertRoles[i] = { roleKey: role, accountUserId: userId };
+  });
+
+  try {
+    // fügt neue Rollen (insertRoles) in die Datenbank Tabelle RolesToAcconts ein
+    if (insertRoles.length > 0) {
+      await db
+        .insert(rolesToAccounts)
+        .values(insertRoles)
+        .onConflictDoNothing();
+    }
+
+    // entfernt Rollen (rolesToRemove) aus der Datenbank Tabelle RolesToAccounts
+    if (rolesToRemove.length > 0) {
+      await db
+        .delete(rolesToAccounts)
+        .where(
+          and(
+            inArray(rolesToAccounts.roleKey, rolesToRemove),
+            eq(rolesToAccounts.accountUserId, userId)
+          )
+        );
+    }
+    revalidatePath("admin/user");
+  } catch (error) {
+    handleNeonError(error);
+  }
+}
 
 export const getRole = cache(async (recordId: number) => {
   try {
